mod: skip when in spectator mode
When in spectator mode, the player character position is the camera's
and if the camera is far in space, where no asteroids are loaded,
this ends up deleting all asteroids where the player character is.
So just skip the asteroid deletion logic when using spectator mode,
only do it when the admin is back in their character.