entity tracking entirely in the mod, using events
Previously, entity tracking was done with a periodic query of all
entities within view distance of the player, sending that to the
server, who would query the same and respond with the valid
entities, as part of the plugin.

Now, the plugin does not do any communication. The mod handles
both client/server, and tracks entities using OnEntityAdd/Remove
events. The plugin patch just ensures the entities we do not want
(i.e. as per config) are not created in the first place.
That effectively makes the mod a generic asteroid entity sync.

Thanks to Echtros for the inspiration.